third_party/WebKit/LayoutTests/fast/layout/common-ancestor-relayout-boundary-expected.txt

PASS window.internals.needsLayoutCount() is 8
PASS window.internals.needsLayoutCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
